Reorganize, add kiwi info
