+++ src/app/models/constants.ts
@@ -1,4 +1,4 @@
-import { Rational } from './math/rational';
+import { Rational } from './rational';
 
 export const WAGON_STACKS = new Rational(BigInt(40));
 export const WAGON_FLUID = new Rational(BigInt(25000));
