+++ src/app/models/data/belt.ts
@@ -1,4 +1,4 @@
-import { Rational } from '../math/rational';
+import { Rational } from '../rational';
 
 export interface Belt {
   speed: number;
