+++ src/app/models/data/module.ts
@@ -1,4 +1,4 @@
-import { Rational } from '../math/rational';
+import { Rational } from '../rational';
 
 export interface Module {
   speed?: number;
