+++ src/app/models/data/recipe.ts
@@ -1,4 +1,4 @@
-import { Rational } from '../math/rational';
+import { Rational } from '../rational';
 import { Entities } from '../entities';
 import { ItemId } from './item';
