+++ src/app/models/enum/display-rate.ts
@@ -1,4 +1,4 @@
-import { Rational } from '../math/rational';
+import { Rational } from '../rational';
 
 export enum DisplayRate {
   PerSecond = 1,
