+++ src/app/models/enum/research-speed.ts
@@ -1,4 +1,4 @@
-import { Rational } from '../math/rational';
+import { Rational } from '../rational';
 
 export enum ResearchSpeed {
   Speed0 = 0,
