--- src/app/models/math/index.ts
@@ -1,6 +1,0 @@
-export * from './constraint';
-export * from './expression';
-export * from './rational';
-export * from './solver';
-export * from './strength';
-export * from './variable';
