+++ src/app/models/step.ts
@@ -1,7 +1,7 @@
 import { ItemId } from './data/item';
 import { RecipeId } from './data/recipe';
 import { Entities } from './entities';
-import { Rational } from './math/rational';
+import { Rational } from './rational';
 
 export interface Step {
   itemId: ItemId;
